requirements.txt
appnope==0.1.4
asttokens==3.0.0
black==23.11.0
cachetools==5.5.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
coverage==7.11.0
debugpy==1.8.17
decorator==5.2.1
dotenv==0.9.9
executing==2.2.1
flake8==6.1.0
google-api-core==2.27.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-bigquery==3.13.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.62.3
httplib2==0.31.0
idna==3.11
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.9.1
matplotlib-inline==0.2.1
mccabe==0.7.0
mypy==1.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==1.24.3
oauthlib==3.3.1
packaging==25.0
pandas==2.1.3
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.19.2
pyparsing==3.2.5
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.1
pyzmq==27.1.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
stack-data==0.6.3
structlog==23.2.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
apify-client==1.6.0